TML Estimator for the Natural Direct Effect

O = (W, A, Z, Y) W = Covariates (possibly multivariate) A = Treatment (binary or categorical) Z = Mediators (binary or categorical; possibly multivariate) Y = Outcome (binary or bounded continuous)

tmle_NDE(e_learners, psi_Z_learners, max_iter = 10000, step_size = 1e-06, ...)

Arguments

e_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in fitting a cleverly parameterized propensity score that conditions on the mediators, i.e., \(e = P(A \mid Z, W)\).

psi_Z_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in a regression of a pseudo-outcome on the baseline covariates, i.e., \(psi_Z(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) \mid A = 0, W]\).

max_iter
A numeric setting the maximum iterations allowed in the targeting step based on universal least favorable submodels.

step_size
A numeric giving the step size (delta_epsilon in tmle3) to be used in the targeting step based on universal least favorable submodels.

...
Additional arguments (currently unused).